Modularize tests
