+++ tests/read/test_utils.py
@@ -2,7 +2,7 @@ import pytest
 
 from cvrplib.read.utils import find_set
 
-from ._utils import selected_cases
+from .._utils import selected_cases
 
 
 @pytest.mark.parametrize("case", selected_cases())
